+++ pkg/metadata/load.go
@@ -2,6 +2,7 @@ package metadata
 
 import (
 	"encoding/json"
+	apexLog "github.com/apex/log"
 	"io/ioutil"
 )
 
@@ -13,5 +14,6 @@ func (tm *TableMetadata) Load(location string) (uint64, error) {
 	if err := json.Unmarshal(data, tm); err != nil {
 		return 0, err
 	}
+	apexLog.Debugf("success TableMedata.Load(%s)", location)
 	return uint64(len(data)), nil
 }